+++ src/examples/hello_world/mod.rs
@@ -1,7 +1,17 @@
 use std::fs;
 use std::path::Path;
 
+use crate::metadata::*;
 use crate::utils::to_crate_name;
+
+// metadata files
+
+pub fn config_metadata(path: &Path) {
+  // .gitignore
+  let gitignore_content = r#"/target
+    "#;
+  create_gitignore(path, gitignore_content);
+}
 
 pub fn hello_world_cargo_toml(name: &str, fluxor_version:  &str) -> String {
     let crate_name = to_crate_name(name);