+++ src/examples/mod.rs
@@ -1,3 +1,4 @@
 pub mod hello_world;
 pub mod routes;
 pub mod assets;
+pub mod dotenv;