+++ src/examples/routes/mod.rs
@@ -1,7 +1,17 @@
 use std::fs;
 use std::path::Path;
 
+use crate::metadata::*;
 use crate::utils::to_crate_name;
+
+// metadata files
+
+pub fn config_metadata(path: &Path) {
+  // .gitignore
+  let gitignore_content = r#"/target
+    "#;
+  create_gitignore(path, gitignore_content);
+}
 
 // Cargo.toml